Each form you build will have to subclass FieldsController.
This will drive the display of the form. So far, it handles keyboard show/hide and making sure fields stay accessible and visible.

It also handles UITextField keyboard resign with Return key. It should similarly do the right thing for other type of inout fields (like showing done button for UITextView or something like it).