Mockup scenario where you need to insert / delete sections.
This proves to be very tricky for the seamless self-sizing layout, as cached layout Info for items in, say deleted section, will now be wrongly applied to items from remaining subsequent sections.
Same with the insertion.

Thus our custom layout needs to re-shuffle its cached LayoutStore, which is anything but obvious.