Adjusting cachedStore in prepare(forCollectionViewUpdates) is too late.
UICV will call invalidateLayout, then prepare and then layoutAttributesForElements(in:) before that method.

Because, of that, rebuild must happen on invalidateDataSourceCounts otherwise you get lots of: "Layout attributes ... were received from the layout ... but are not valid for the data source counts. Attributes will be ignored." warnings when deleting items.